requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
Faker==37.1.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.0.0
narwhals==1.37.1
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
plotly==6.0.1
propcache==0.2.0
pycparser==2.22
pycryptodomex==3.21.0
Pygments==2.19.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.40
steam==1.4.4
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.2
vdf==3.4
watchdog==6.0.0
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.16.0
